requirements.txt
bentoml
click
joblib
lxml
matplotlib
mlflow
numpy
pandas
Pillow
Pillow
python-dotenv
rich
scikit_learn
scipy
SQLAlchemy
statsmodels
typing_extension
zenml
evidently
